Commands: move `PluginDelegate` to a separate file
